[NFC][MemProf] Move Radix tree methods to their own header and cpp.
