+++ llvm/include/llvm/ProfileData/MemProfReader.h
@@ -22,8 +22,8 @@
 #include "llvm/Object/Binary.h"
 #include "llvm/Object/ObjectFile.h"
 #include "llvm/ProfileData/InstrProfReader.h"
-#include "llvm/ProfileData/MemProf.h"
 #include "llvm/ProfileData/MemProfData.inc"
+#include "llvm/ProfileData/MemProfRadixTree.h"
 #include "llvm/Support/Error.h"
 #include "llvm/Support/MemoryBuffer.h"
